requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Integration Dependencies
openai==1.3.7
google-generativeai==0.3.2
httpx==0.25.2



# Web3 Dependencies
web3==6.12.0
eth-account==0.9.0
eth-utils==2.3.1

# Database and Caching
sqlalchemy==2.0.23

# Security and Authentication
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Graph Processing Dependencies
networkx==3.2.1
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
requests==2.31.0